+++ src/main/java/com/lockers/outerpark/common/security/handler/CustomAuthenticationEntryPoint.java
@@ -1,0 +1,34 @@
+package com.lockers.outerpark.common.security.handler;
+
+import java.io.IOException;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+	private final ObjectMapper objectMapper;
+
+	@Override
+	public void commence(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException authException) throws IOException {
+		log.error("[CustomAuthenticationEntryPoint : commence] 인증 실패로 response.sendError 발생.");
+
+		response.setStatus(401);
+		response.setContentType("application/json");
+		response.setCharacterEncoding("utf-8");
+		response.getWriter().write(objectMapper.writeValueAsString("인증이 실패하였습니다."));
+	}
+
+}
